refactor: share the beforeunload leave guard across session pages
